Add test for default focus out
